backend/requirements.txt
Django==2.0.2
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.0
pytz==2018.7
